Time: 110 ms (56.18%), Space: 198.8 MB (20.42%) - LeetHub

+++ 3453-separate-squares-i/3453-separate-squares-i.cpp
@@ -3,8 +3,8 @@ public:
     double separateSquares(vector<vector<int>>& squares) {
         int maxy = 0;
         int miny = 1e9;
-        long double upper = 0;
-        long double lower = 0;
+        double upper = 0;
+        double lower = 0;
         for(auto& square: squares) {
             miny = square[1] < miny ? square[1] : miny;
             maxy = (square[1] + square[2]) > maxy ? (square[1] + square[2]) : maxy;
